+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Percursus/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Percursus/releases)
 
-- Added: TOC version for patch 11.2.0
+- Added: New Race Tracker background 'N'Zoth'